add aggregation, update readme.md
